This is test code!!
I have changed the declarations of the Rig class functions we use:

open()
setConf()
getFreq()
setFreq()
setPTT()
close()

... so that each one returns the (negative) error code returned by 
hamlib.  These functions no longer throw exceptions.

Initial tests seem to show things working well.  I have found one
condition that causes a crash.  The rig is initially opened and is
running OK (in this case, with the Polling interval set to 1 s).  Then
the radio is turned off.  The next call to getFreq() never returns.

Probably we need to set a timeout limit?

The code is kinda on the quick-and-dirty side.  Improvements will be 
welcom!



git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@3246 ab8295b8-cf94-4d9e-aec4-7959e3be5d79