Added tuning parameters for audio subsystem.
By adding the following section to the initialisation file the audio buffer
sizes and audio thread priority may be adjusted.

[Tune]
Audio\InputBufferFrames=1200
Audio\OutputBufferMs=1000
Audio\ThreadPriority=4

The values above are the program defaults that will be used if the
initialisation parameters are omitted.

Thread prioritis are the QThread::Priority enumumeration values.



git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@3576 ab8295b8-cf94-4d9e-aec4-7959e3be5d79